pyproject.toml
[tool.poetry]
name = "graphql-next-ws"
version = "1.0.0"
description = ""
authors = ["Devin Fee <devin@devinfee.com>"]
packages = [
    { include = "graphql_ws" }
]

[tool.poetry.dependencies]
python = "^3.6"
aiohttp = "^3.4"
graphql-core-next = "^1.0"


[tool.poetry.dev-dependencies]
black = "^18.3-alpha.0"
ipdb = {version = "^0.11.0", optional = true}
mypy = "^0.620.0"
pylint = "^2.1"
pytest = "^3.0"
pytest-asyncio = "^0.9.0"

[tool.poetry.extras]
dev = ["ipdb"]

[tool.black]
line-length = 80
